intcode: Fix case where the last changed rel base wasn't saved
This case would happen when in an intcode operation session (period
between receiving the input, until the next input), no rel base change
opCode would occur.

The bug was that the relative base was saved to the state only when
that opcode was called.

It should have been (as is now) saved after any opcode or before any
return from the intcode `process` proc.